move view_utils
